Change layout
Re-structured the quesadilla page a bit; added
a few divs and sections and played around with
margin and padding to see how that changes the
layout. Basically for practice.

However, I did somehow introduce a bug. The 2nd
div with the class div-half is supposed to be the
second container for the ingredients and steps,
and should be aligned with the first container on
top. However, that is not the case. There is a lot
of empty space between the headline and the 2nd
div half container. I do not know why - yet.